+++ core/indexing/docs/article.ts
@@ -2,7 +2,7 @@ import { Readability } from "@mozilla/readability";
 import { JSDOM } from "jsdom";
 import { Chunk } from "../../";
 import { cleanFragment, cleanHeader } from "../chunk/markdown";
-import { PageData } from "./crawlSite";
+import { PageData } from "./crawl";
 
 export type ArticleComponent = {
   title: string;
